+++ packages/furo-data-input/demo/demo-furo-data-bool-icon.js
@@ -4,12 +4,12 @@ import {FBP} from "@furo/fbp";
 import "@furo/doc-helper"
 import "../furo-catalog"
 /**
- * `demo-furo-bool-icon`
+ * `demo-furo-data-bool-icon`
  *
  * @customElement
  * @appliesMixin FBP
  */
-class DemoFuroBoolIcon extends FBP(LitElement) {
+class DemoFuroDataBoolIcon extends FBP(LitElement) {
 
   /**
    * Themable Styles
@@ -40,15 +40,15 @@ class DemoFuroBoolIcon extends FBP(LitElement) {
   render() {
     // language=HTML
     return html`
-      <h2>Demo demo-furo-bool-icon</h2>
+      <h2>Demo demo-furo-data-bool-icon</h2>
       <p>description</p>
       <furo-demo-snippet >
         <template>
-          <furo-bool-icon ></furo-bool-icon>
+          <furo-data-bool-icon ></furo-data-bool-icon>
         </template>
       </furo-demo-snippet>
     `;
   }
 }
 
-window.customElements.define('demo-furo-bool-icon', DemoFuroBoolIcon );
+window.customElements.define('demo-furo-data-bool-icon', DemoFuroDataBoolIcon );