+++ packages/furo-navigation/stock/furo-recursive-tree-b.js
@@ -2,7 +2,7 @@ import {LitElement, html, css} from 'lit-element';
 import {Theme} from "@furo/framework/theme"
 import {FBP} from "@furo/fbp";
 import "@furo/fbp/flow-repeat"
-import "@furo/data-input/furo-bool-icon";
+import "@furo/data-input/furo-data-bool-icon";
 import "@furo/layout/furo-horizontal-flex";
 import {FuroRecursiveTree} from "./furo-recursive-tree"
 
@@ -27,7 +27,7 @@ class FuroRecursiveTreeB extends FuroRecursiveTree {
 
   <div class="label" @-click="--labelClicked" @-dblclick=":STOP, --dblclicked">   
   <span class="oc">
-    <furo-bool-icon ?hidden="${!this.field.children.repeats.length}" ƒ-toggle="--dblclicked" ƒ-bind-data="--fieldOpen"></furo-bool-icon>
+    <furo-data-bool-icon ?hidden="${!this.field.children.repeats.length}" ƒ-toggle="--dblclicked" ƒ-bind-data="--fieldOpen"></furo-data-bool-icon>
     </span>       
     <span class="name">${this.field.display_name}</span>          
     <span class="desc">${this.field.description}</span>